Add files via upload
Project Title: Netflix Clone Description: Welcome to my Netflix Clone project! 🎬🍿 This is a web application built using HTML and CSS that emulates the look and feel of the popular streaming platform, Netflix. With a responsive design and carefully crafted UI, this clone offers users a familiar and engaging streaming experience.